… test/optimizations/doNoWork.js