This attempts to solve #20 with the following optimizations:

- In any chain of imlazy iterables it reduces the number of iterators created to just 2 - one from the original iterable and one from the final iterable. It does this by storing operations and iterables on the lazy iterables themselves
- It checks whether the iterable provided is an array and if so iterates with a standard for loop instead of a for...of loop

Performance in the benchmark over infinite iterables improves from around 391 ops/sec to 636 ops/sec and over arrays from around 3,805 ops/sec to 7,618 ops/sec

But is still far short of transducers - about 2.2x faster on infinite iterables and 3.0x faster on arrays

One disadvantage of this approach is that the `doNoWork` optimization which sends a message up the chain not to operate on unwanted data is no longer implemented - but this optimization is only ever useful in cases where the user removes items from an iterable after applying heavy mapping functions which seems a pretty niche use-case. It was undocumented anyway so I'm not too worried about losing it