@benji6 2017-06-17
As the benchmarks have been significantly improved I'm going to merge this, but I still think there are things to think about in this area